Add 'component' argument to build single component
By default, all components are built. This could be expensive if there
are many components in a WMT instance. The 'components' argument
specifies a single component to build; all other components are
ignored.

Note that the 'component' argument gets translated to the 'single_component'
argument internally to prevent a variable collision. The collection of
functions in this module should be rewritten as a class.